# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Form-Builder SaaS API",
        "version": "1.0",
        "description": "The versioned REST API for the Meridian form-builder platform. Served per tenant on the tenant subdomain (https://{tenant}.meridian.test/api/v1) and authenticated with Sanctum personal-access-token API keys. See docs/api-specification.md."
    },
    "servers": [
        {
            "url": "https://{tenant}.meridian.test/api/v1",
            "description": "Per-tenant API base \u2014 replace {tenant} with your workspace subdomain slug.",
            "variables": {
                "tenant": {
                    "default": "acme"
                }
            }
        }
    ],
    "security": [
        {
            "sanctumToken": []
        }
    ],
    "paths": {
...
        "/sync/submissions": {
            "post": {
                "operationId": "v1.sync.submissions",
                "summary": "Replay a batch of queued offline submissions (idempotent; per-item results)",
                "tags": [
                    "SyncSubmission"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncSubmissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SyncSubmissionResultResource"
                                            }
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumToken": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "SyncSubmissionRequest": {
                "type": "object",
                "description": "The offline batch-replay request (Increment G8b, docs/offline-first-sync-design.md \u00a75). Authorization is\nthe route's `ability:write:submissions` middleware + RLS. Only the coarse shape is asserted here \u2014 each\nitem's per-field validation is the shared {@see SubmissionPipeline}, one item at\na time, so a partial failure never rolls back its siblings. `client_submission_uuid` is the idempotency key\n(a duplicate replay is a 200 no-op); `submitted_at` is accepted but informational \u2014 the server timestamp is\nauthoritative (architecture \u00a74.2). Bounded to 100 items per request.",
                "properties": {
                    "submissions": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "form_version_id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "client_submission_uuid": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "answers": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "locale": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "maxLength": 10
                                },
                                "device_id": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "maxLength": 100
                                },
                                "app_version": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "maxLength": 20
                                },
                                "submitted_at": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "format": "date-time"
                                }
                            },
                            "required": [
                                "form_version_id",
                                "client_submission_uuid",
                                "answers"
                            ]
                        },
                        "maxItems": 100
                    }
                },
                "required": [
                    "submissions"
                ],
                "title": "SyncSubmissionRequest"
            },
            "SyncSubmissionResultResource": {
                "type": "object",
                "properties": {
                    "client_submission_uuid": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "submission": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "client_submission_uuid",
                    "status",
                    "submission",
                    "error"
                ],
                "title": "SyncSubmissionResultResource"
            },
...
        },
        "responses": {
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}